Change the playfield tile graphics
Change the bg tiles for the playfield such that uncleared and cleared tiles are both a single solid color - gray or black. This gives the playfield a much flatter character which I'll try out for a while.

Also allows us to use different colors for each in-progress line segment so changed those to be red/blue instead of both being orange. These colors are controlled via palette, could be changed per-level.

Reorganized the HUD in the playfield screen a little bit to make it more pleasant looking and changed the color of the text to near-white.

Added support to disable drawing the nearest tile highlight under the player and disabled it for now. This looks clean but imprecise.