Improve update_cleared_playfield_tiles performance (#61)
Re-implements update_cleared_playfield_tiles() to walk over the playfield with a single pointer instead of keeping indices. Also optimizes the math and logic as much as I think is possible while keeping it only in C.

This brings the entire sweep from 141747 to 113467 cycles (~20% or ~30000 fewer cycles) in some test scenario. This probably saves a frame or two during the sweep.

Performance while iterating over a completely unmarked playfield improves by about 10%.

Performance while iterating over a playfield with 44 marked tiles improves by about 27%.

The more tiles which are marked, the more noticeable the improvement would be.